Menu: only expand active hierarchy
Before this change, the entire menu tree below the selected top-level heading
was expanded.

This change adds some triangles on menu items at level 2 and below to indicate
which ones can be or are expanded. I didn't put them on the top level, because
I felt that it looked too cluttered and didn't add much value.

Fixes #88